This commit introduces laziness to the Plack-Debugger UI.
We only load the initial AJAX data so that we can know if
we need to do things like; track AJAX requests, open the
toolbar if there is an error or warning, etc. We do not
attempt to render any buttons or panels at this time, we
only stash the data. We do create empty DOM nodes for
the buttons/panels though, this is because we rely on
the index of the button DOM node being the same as the
index of the Panel DOM node (NOTE: we should fix this).

When the toolbar is expanded we render all the buttons,
but nothing more. When a particular panel is opened, we
render that panel, but nothing more.

If AJAX tracking is enabled, we also try to be lazy about
this as well. If the button and/or panel is already been
rendered, we will just update it. If the button and/or
panel has not been rendered, we merge the new AJAX data
into the existing data and let the normal lazy-render
process happen.

There is still room for improvement here, for instance
we are still (re)loading all the AJAX data on every
AJAX request, and still creating empty DOM nodes, but
this should be a great improvement over the previous
version.

All in good time, all in good time.